terminer un premier mapping fonctionnel

+++ ErabliereApi/Services/Weather/AccuWeatherModels/WeatherForecastResponse.cs
@@ -25,7 +25,7 @@ public class Headline
 public class Dailyforecast
 {
     public DateTimeOffset? Date { get; set; }
-    public int EpochDate { get; set; }
+    public long EpochDate { get; set; }
     public Temperature? Temperature { get; set; }
     public Day? Day { get; set; }
     public Night? Night { get; set; }
